[Plugin] Add basic resolution of `ExternalReference` `target_url` through OpenAssetIO
Adds simple resolution of entity references set in the target_url
property of an OTIO ExternalReference through OpenAssetIO.

This currently requires that the media linker is configured via the
media_linker_argument_map provided to otio.adapters.read_from_string.

There are a variety of options for how we could improve this over time
(through clip + inherited metadata, env var to settings file, etc.), we
just wanted to get things going at this point to better understand
the rest of the implementation.

The additional steps required to grab the OpenAssetIO dependencies are
a little bit cellotape and string at this point. We need to properly
sort out how we want to manage the release of example code that is part
of the main OpenAssetIO project.

Part of #215.