[Plugin] Switch to the new unified entity data model
This now requires the use of the `feature/181-coreCppBuild` branch
of OpenAssetIO.

As we’re still early doors with this, we’re happy that the process of
getting all this running is still somewhat manual. Once we have PyPy
packages published for `openassetio` and `openassetio_mediacreation`
this should get a lot simpler.

Signed-off-by: Tom Cowland <tom@foundry.com>