[Build] Bump OpenAssetIO dependencies
Part of OpenAssetIO/OpenAssetIO#1311. Deprecated aliases have been
removed from the latest OpenAssetIO, but the version of MediaCreation
that was pinned in `setup.py` still used them.

This wasn't actually a problem, since the version of OpenAssetIO was
also pinned.

However, it is a useful canary if we only specify a minimum range for
downstream projects, whenever possible. Upon the next OpenAssetIO
release we should re-check this project (and others) for compatibility.

So remove the maximum bound for OpenAssetIO, bump the minimum bound
for OpenAssetIO-MediaCreation, and bump the minimum BAL version used for
testing, such that it supports the later OpenAssetIO versions.

Signed-off-by: David Feltell <david.feltell@foundry.com>